+++ packages/generate-design-tokens/index.js
@@ -160,6 +160,9 @@ export const get${toPascalCase(skinName)}Skin: GetKnownSkin = () => {
                 )
                 .join(',')},
         },
+        spacing: ${JSON.stringify(
+            Object.fromEntries(Object.entries(designTokens.spacing).map(([name, {value}]) => [name, value]))
+        )},
     };
     return skin;
 };