+++ src/__screenshot_tests__/navigation-bar-screenshot-test.tsx
@@ -88,6 +88,42 @@ test('MainNavigationBar wide', async () => {
     expect(await page.screenshot()).toMatchImageSnapshot();
 });
 
+test('MainNavigationBar desktop with topSlot', async () => {
+    const page = await openStoryPage({
+        id: 'components-navigation-bars-mainnavigationbar--default',
+        device: 'DESKTOP',
+        args: {topSlot: true, desktopLargeMenu: true, menu: 'default'},
+    });
+
+    expect(await page.screenshot()).toMatchImageSnapshot();
+
+    await page.click(await screen.findByRole('button', {name: 'Start'}));
+    expect(await page.screenshot()).toMatchImageSnapshot();
+});
+
+test('MainNavigationBar mobile with topSlot', async () => {
+    const page = await openStoryPage({
+        id: 'components-navigation-bars-mainnavigationbar--default',
+        device: 'MOBILE_IOS',
+        args: {topSlot: true},
+    });
+
+    expect(await page.screenshot()).toMatchImageSnapshot();
+
+    await page.click(await screen.findByRole('button', {name: 'Abrir menú de navegación'}));
+    expect(await page.screenshot()).toMatchImageSnapshot();
+});
+
+test('MainNavigationBar desktop with topSlot and custom background', async () => {
+    const page = await openStoryPage({
+        id: 'components-navigation-bars-mainnavigationbar--default',
+        device: 'DESKTOP',
+        args: {topSlot: true, topSlotBackgroundColor: '#FABADA'},
+    });
+
+    expect(await page.screenshot()).toMatchImageSnapshot();
+});
+
 test.each`
     variant          | isDarkMode | device
     ${'default'}     | ${false}   | ${'DESKTOP'}
@@ -245,7 +281,7 @@ test.each(['large' /* 'small' */])('MainNavigationBar inverse with %s menu in DE
         args: {sections: true, desktopLargeMenu: menuType === 'large', menu: 'default', variant: 'brand'},
     });
 
-    // first section opened
-    await page.click(await screen.findByRole('button', {name: 'Start'}));
+    const startButton = await screen.findByRole('button', {name: 'Start'});
+    await page.click(startButton);
     expect(await page.screenshot()).toMatchImageSnapshot({failureThreshold: 0.00002});
 });