+++ src/__stories__/card-cover-story.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import {CoverCard, Placeholder} from '..';
 import {
     commonArgTypes,
+    dataArgTypes,
     defaultCommonCardArgs,
     getAsset,
     getButtonLink,
@@ -78,7 +79,7 @@ export const Default: StoryComponent<CoverCardArgs> = ({
                 videoSrc={videoNameToUrl[videoSrc as never]}
                 videoLoop={videoLoop}
                 videoAutoPlay={videoAutoPlay}
-                slot={slot ? <Placeholder height={50} /> : undefined}
+                slot={slot ? [<Placeholder height={30} />, <Placeholder height={30} />] : undefined}
                 footerSlot={footerSlot ? <Placeholder height={50} /> : undefined}
                 buttonPrimary={getButtonPrimary(buttonPrimary)}
                 buttonSecondary={getButtonSecondary(buttonSecondary)}
@@ -109,4 +110,5 @@ Default.args = {
 Default.argTypes = {
     ...commonArgTypes,
     ...imageAndVideoArgTypes,
+    ...dataArgTypes,
 };