+++ src/select.css.ts
@@ -10,7 +10,6 @@ import {
     iconButtonSize,
 } from './text-field-base.css';
 import {pxToRem} from './utils/css';
-import {browserDefaultFocusOutline} from './text-field-components.css';
 
 const top = createVar();
 const left = createVar();
@@ -27,6 +26,7 @@ export const vars = {
 };
 
 export const chevronSize = pxToRem(20);
+export const optionContainerPadding = 8;
 
 export const optionsContainer = style([
     sprinkles({
@@ -41,10 +41,8 @@ export const optionsContainer = style([
         top,
         left,
         minWidth,
-        boxShadow:
-            '0px 5px 5px -3px rgba(0,0,0,0.2), 0px 8px 10px 1px rgba(0,0,0,0.14), 0px 3px 14px 2px rgba(0,0,0,0.12)',
-        paddingTop: 8,
-        paddingBottom: 8,
+        boxShadow: '0px 2px 4px 0px #00000033',
+        padding: optionContainerPadding,
         transformOrigin,
         transition: 'opacity .03s linear,transform .12s cubic-bezier(0,0,.2,1) .15s',
         overflowY: 'auto',
@@ -68,16 +66,11 @@ const selectContainerBase = style([
         cursor: 'pointer',
     }),
     {
-        borderRadius: `calc(${skinVars.borderRadii.input} - 1px)`,
+        outline: 'none',
         '@media': {
             [mq.tabletOrSmaller]: {
                 width: '100%',
             },
-        },
-        ':focus': {
-            outline: browserDefaultFocusOutline,
-            border: `1px solid transparent`, // needed for outline to follow border radius in Safari
-            margin: -1, // to avoid layout shift when adding the transparent border
         },
     },
 ]);
@@ -198,7 +191,7 @@ export const iconContainer = style({
     alignItems: 'center',
     justifyContent: 'center',
     // remove extra button space on the right so that icon is not too far from field's container
-    marginRight: -12,
+    marginRight: -14,
 });
 
 export const menuItem = style([
@@ -208,8 +201,12 @@ export const menuItem = style([
         alignItems: 'center',
         cursor: 'pointer',
         paddingY: 8,
-        paddingX: 16,
+        paddingX: 8,
     }),
+    {
+        borderRadius: `calc(${skinVars.borderRadii.popup} - ${optionContainerPadding / 2}px)`,
+    },
+
     {
         height: pxToRem(48),
         transition: 'background-color 150ms cubic-bezier(0.4, 0, 0.2, 1) 0ms',
@@ -217,7 +214,7 @@ export const menuItem = style([
         '@media': {
             [mq.supportsHover]: {
                 ':hover': {
-                    backgroundColor: 'rgba(0, 0, 0, 0.08)',
+                    backgroundColor: skinVars.colors.backgroundContainerHover,
                 },
             },
         },