+++ src/skins/types/colors.tsx
@@ -24,6 +24,14 @@ export type Colors = {
     backgroundBrandBottom: string;
     appBarBackground: string;
     navigationBarBackground: string;
+    backgroundDropZoneHover: string;
+    backgroundDropZoneDragover: string;
+    backgroundDropZoneBrandHover: string;
+    backgroundDropZoneBrandDragover: string;
+    backgroundDropZoneNegativeHover: string;
+    backgroundDropZoneNegativeDragover: string;
+    backgroundDropZoneMediaHover: string;
+    backgroundDropZoneMediaDragover: string;
     skeletonWave: string;
     borderLow: string;
     border: string;
@@ -214,6 +222,7 @@ export type Colors = {
     textNavigationSearchBarText: string;
     textAppBar: string;
     textAppBarSelected: string;
+    iosGlassAppBar: string;
     iosGlassAppBarSelected: string;
     customTabsBackground: string;
     tagTextPromo: string;