+++ src/text-field-base.css.ts
@@ -8,7 +8,8 @@ import {pxToRem} from './utils/css';
 const borderSize = 1;
 
 // We need to substract border size from padding because the container has boxSizing: border-box
-export const fieldVerticalPadding = 8 - borderSize;
+export const fieldTopPadding = `calc(${vars.spacing.inputPadding.top} - ${borderSize}px)`;
+export const fieldBottomPadding = `calc(${vars.spacing.inputPadding.bottom} - ${borderSize}px)`;
 export const fieldLeftPadding = 12 - borderSize;
 export const fieldRightPadding = 16 - borderSize;
 
@@ -34,6 +35,8 @@ const helperTextDesktopLineHeight = createVar();
 const labelScaleDesktop = createVar();
 const labelScaleMobile = createVar();
 
+export const suggestionsContainerPadding = 8;
+
 export const fieldVars = {
     mobileFontSize,
     desktopFontSize,
@@ -52,23 +55,23 @@ export const fieldVars = {
 };
 
 const topSpaceWithLabel = {
-    desktop: `calc(${shrinkedLabelDesktopLineHeight} + ${fieldVerticalPadding}px)`,
-    mobile: `calc(${shrinkedLabelMobileLineHeight} + ${fieldVerticalPadding}px)`,
+    desktop: `calc(${shrinkedLabelDesktopLineHeight} + ${fieldTopPadding})`,
+    mobile: `calc(${shrinkedLabelMobileLineHeight} + ${fieldTopPadding})`,
 };
 
 const topSpaceWithoutLabel = {
-    desktop: `calc(${shrinkedLabelDesktopLineHeight} / 2 + ${fieldVerticalPadding}px)`,
-    mobile: `calc(${shrinkedLabelMobileLineHeight} / 2 + ${fieldVerticalPadding}px)`,
+    desktop: `calc(${shrinkedLabelDesktopLineHeight} / 2 + ${fieldTopPadding})`,
+    mobile: `calc(${shrinkedLabelMobileLineHeight} / 2 + ${fieldTopPadding})`,
 };
 
 const bottomSpaceWithLabel = {
-    desktop: fieldVerticalPadding,
-    mobile: fieldVerticalPadding,
+    desktop: fieldBottomPadding,
+    mobile: fieldBottomPadding,
 };
 
 const bottomSpaceWithoutLabel = {
-    desktop: `calc(${shrinkedLabelDesktopLineHeight} / 2 + ${fieldVerticalPadding}px)`,
-    mobile: `calc(${shrinkedLabelMobileLineHeight} / 2 + ${fieldVerticalPadding}px)`,
+    desktop: `calc(${shrinkedLabelDesktopLineHeight} / 2 + ${fieldBottomPadding})`,
+    mobile: `calc(${shrinkedLabelMobileLineHeight} / 2 + ${fieldBottomPadding})`,
 };
 
 const commonInputStyles = style([
@@ -362,6 +365,7 @@ export const menuItemBase = style([
         minHeight: pxToRem(48),
         padding: '6px 8px',
         userSelect: 'none',
+        borderRadius: `calc(${skinVars.borderRadii.popup} - ${suggestionsContainerPadding / 2}px)`,
     },
 ]);
 
@@ -396,7 +400,7 @@ export const suggestionsContainer = style([
         marginTop: 8,
         boxSizing: 'border-box',
         boxShadow: '0px 2px 4px 0px #00000033',
-        padding: 8,
+        padding: suggestionsContainerPadding,
         background: skinVars.colors.backgroundContainer,
         borderRadius: skinVars.borderRadii.popup,
 
@@ -409,9 +413,4 @@ globalStyle(`${suggestionsContainer} > ul`, {
     listStyleType: 'none',
     padding: 0,
     margin: 0,
-});
-
-export const fieldEndIconContainer = style({
-    // remove extra button space on the right so that icon is not too far from field's container
-    marginRight: -12,
 });